handle unboxing

+++ de.xwic.appkit.core/src/de/xwic/appkit/core/model/util/EntityUtil.java
@@ -56,7 +56,7 @@ public final class EntityUtil {
 	 * @return
 	 */
 	public static <E extends IEntity> E getEntity(final Class<E> entityClass, final Integer id) {
-		return id == null ? null : (E) findDAO(entityClass).getEntity(id);
+		return id == null ? null : (E) findDAO(entityClass).getEntity(id.intValue());
 	}
 
 	/**
@@ -65,7 +65,7 @@ public final class EntityUtil {
 	 * @return
 	 */
 	public static <E extends IEntity> E getOrCreateEntity(final Class<E> entityClass, final Integer id) {
-		if (id == null || id < LOWEST_POSSIBLE_ID) {
+		if (id == null || id.intValue() < LOWEST_POSSIBLE_ID) {
 			return createEntity(entityClass);
 		}
 		return getEntity(entityClass, id);
